requirements.txt
python-telegram-bot==20.7
openai==1.3.5
pillow==10.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
aiofiles==23.2.1
matplotlib==3.8.1
pandas==2.1.3
requests==2.31.0
aiohttp==3.9.0
uvloop==0.19.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
